Doruk - Ekleme / Güncelleme - AddTransactionRequestDto yeni istekler doğrultusunda düzenlendi. - Para birimi eklendiği için Long tutulan Salary/Totalcost/YearsTotal gibi değişkenlerin tipi Double ile değiştirildi. - Database içerisine dosya kaydı yapılabilmesi için FileEntitysi oluşturuldu, katmanları eklendi ve Api gateway erişimi sağlandı. - FinancialTransaction entitysi istekler doğrultusunda düzenlendi. - ITransactionRepository içerisindeki findAllByTransactionDateBetweenAndIsPaidAndTypeAndCompanyIdAndStatus, annualProfitLoss ve findAllExpnses metodları status eklentisi sebebi ile yeniden düzenlendi. - Önyüz için expenses ve currencies enumlarının değerlerini döndüren(expenses "MANAGER" hariç) endpointler hazırlandı. - Yeni eklenen fieldlar Dummy datalara da eklendi. - Currency tipleri, Expense tipleri, Transactiion durumları enum olarak eklendi.

+++ auth-service/src/main/java/com/ikolay/dto/requests/RegisterRequestDto.java
@@ -44,5 +44,5 @@ public class RegisterRequestDto {
     // EMPLOYEE KAYDI YAPARKEN MANAGER'IN COMPANY ID'SI KULLANILACAK.
     private Long companyId;
     // EMPLOYEE KAYDI SIRASINDA ALINACAK
-    private Long salary;
+    private Double salary;
 }
